Implement MIDI 2.0 full precision: 16-bit velocity, 32-bit CC/pitch bend
Replace 7-bit MIDI 1.0 processing with native 16/32-bit MIDI 2.0 precision
throughout the entire pipeline (MIDIEventQueue → MIDIInputManager →
M2DXAudioEngine → FMSynthEngine). MIDI 1.0 input is upscaled at the
MIDIInputManager boundary; UMP data from CoreMIDITransport is decoded
directly at full precision via the new handleUMPData() path.

+++ M2DXPackage/Sources/M2DXFeature/MIDIEventQueue.swift
@@ -15,8 +15,8 @@ struct MIDIEvent: Sendable {
     }
 
     let kind: Kind
-    let data1: UInt8
-    let data2: UInt8
+    let data1: UInt8       // note number or CC number (7-bit)
+    let data2: UInt32      // velocity16, CC value32, or pitchBend32
 }
 
 // MARK: - MIDI Event Queue
